@@ -17,8 +17,8 @@ package com.jahirfiquitiva.chip
 
 import android.content.res.Resources
 import android.graphics.Color
-import android.support.annotation.ColorInt
-import android.support.annotation.FloatRange
+import androidx.annotation.ColorInt
+import androidx.annotation.FloatRange
 
 /**
  * Created by Allan Wang
