@@ -16,8 +16,8 @@
 package com.jahirfiquitiva.chipview
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
 import com.jahirfiquitiva.chip.ChipView
 
 class MainActivity : AppCompatActivity() {
